requirements.txt
ccxt~=4.5.34
pandas~=3.0.0
python-dotenv~=1.2.1